Update speed_dial.dart
Hero tags should be set to null if they are not used. Setting the heroTag to null prevents the FAB from being wrapped in a hero widget. Hero widgets cannot be the descendant of another Hero widget so heroTags should be set null if they are not used, allowing the FAB to have more flexibility.